test: isolate OS-singleton tests into spec/serial/ for post-parallel execution
