+++ spec/api-shell.spec.ts
@@ -96,21 +96,6 @@ describe('shell module', () => {
         requestReceived
       ]);
     });
-
-    ifit(process.platform === 'darwin')(
-      'removes focus from the electron window after opening an external link',
-      async () => {
-        const url = 'http://127.0.0.1';
-        const w = new BrowserWindow({ show: true });
-
-        await once(w, 'focus');
-        expect(w.isFocused()).to.be.true();
-
-        await Promise.all<void>([shell.openExternal(url), once(w, 'blur') as Promise<any>]);
-
-        expect(w.isFocused()).to.be.false();
-      }
-    );
   });
 
   describe('shell.trashItem()', () => {
